requirements.txt
streamlit==1.43.2
requests==2.32.3
Pillow==11.1.0
python-dotenv==1.0.1 